separate signal utils from axis utils
