+++ lumispy/__init__.py
@@ -21,7 +21,7 @@ import logging
 
 _logger = logging.getLogger(__name__)
 
-from lumispy.utils.axes import nm2eV, eV2nm, nm2invcm, invcm2nm, join_spectra
+from lumispy.utils.axes import nm2eV, eV2nm, nm2invcm, invcm2nm
 from lumispy.utils.io import to_array, savetxt
 
 from lumispy import signals
